requirements.txt
streamlit
pymupdf
easyocr
numpy<2
langchain
langchain-community
sentence-transformers
faiss-cpu
groq
python-dotenv
pyyaml
pillow
